+++ src/Sentry/Internal/Hub.cs
@@ -506,6 +506,29 @@ internal class Hub : IHub, IDisposable
         }
     }
 
+    public void CaptureFeedback(SentryFeedback feedback, Action<Scope> configureScope) =>
+        CaptureFeedback(feedback, null, configureScope);
+
+    public void CaptureFeedback(SentryFeedback feedback, SentryHint? hint, Action<Scope> configureScope)
+    {
+        if (!IsEnabled)
+        {
+            return;
+        }
+
+        try
+        {
+            var clonedScope = CurrentScope.Clone();
+            configureScope(clonedScope);
+
+            CaptureFeedback(feedback, clonedScope, hint);
+        }
+        catch (Exception e)
+        {
+            _options.LogError(e, "Failure to capture feedback");
+        }
+    }
+
     public void CaptureFeedback(SentryFeedback feedback, Scope? scope = null, SentryHint? hint = null)
     {
         if (!IsEnabled)
@@ -513,8 +536,15 @@ internal class Hub : IHub, IDisposable
             return;
         }
 
-        scope ??= CurrentScope;
-        CurrentClient.CaptureFeedback(feedback, scope, hint);
+        try
+        {
+            scope ??= CurrentScope;
+            CurrentClient.CaptureFeedback(feedback, scope, hint);
+        }
+        catch (Exception e)
+        {
+            _options.LogError(e, "Failure to capture feedback");
+        }
     }
 
 #if MEMORY_DUMP_SUPPORTED